Addresses issue #176, "Save" button now behaves as "Save As" button when there is no protocol file currently open.
Fixed name conflict. Function on_selection(), found in the main code, is now named on_selection_function(). This avoids conflict with the on_selection method of Kivy FileChooser.

Preexisting feature: when closing LumaViewPro (LVP), the file name of the protocol last used is stored in a cache file data/current.json. New feature in this commit: when opening LVP, if it is not possible to load the previous protocol, the file name is cleared from memory as well. (Otherwise the "save" button will have the undesired behavior of writing to a file with the old name, rather than letting one name a new file.)